[EDIT]: Moved Node struct to new hpp file
